moved onSplitsArrivedFactory logic to partial producer
